// src/app/modules/campaign-management/models/response/SearchBroadcastModel.ts
export interface SearchBroadcastModel {
    broadcastConfigurationId?: number;
    broadcastId?:number
    broadcastName?: string
    messageTypeName?: string
    broadcastDate?: string
    statusName?: string
    recipient?: number
    delivered?: number
    createdDate?: string
    createdBy?: string
    updatedBy?: string
    updatedDate?: string
    
}